moved voting feature in live_room_page.dart to voting.dart
