+++ lib/live_room_page.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:spotify_polls/song_cards.dart';
+import 'package:spotify_polls/voting.dart';
 
 class LiveRoomPage extends StatefulWidget {
   const LiveRoomPage({super.key, this.title = "Live Room Page"});
@@ -13,21 +14,7 @@ class LiveRoomPage extends StatefulWidget {
 class _LiveRoomPageState extends State<LiveRoomPage> {
   bool playState = false;
 
-  final List<SongCardData> _songCards = [];
 
-  void _addSong() {
-    setState(() {
-      _songCards.insert(
-          0,
-          SongCardData(
-            songName: "Song ${_songCards.length + 1}",
-            artistName: "Artist ${_songCards.length + 1}",
-            trackArt: Image.network(
-                'assets/trackArtPlaceholder.png'),
-            votes: [0, 0],// Placeholder art
-          ));
-    });
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -48,13 +35,6 @@ class _LiveRoomPageState extends State<LiveRoomPage> {
       });
     }
 
-    void voteYes() {
-      _songCards[_songCards.length - 1].votes[0] += 1;
-    }
-    void voteNo() {
-      _songCards[_songCards.length - 1].votes[1] += 1;
-    }
-
     ButtonStyle playerButtonStyle = ElevatedButton.styleFrom(
         padding: EdgeInsets.all(containerHeight * 0.01),
         minimumSize: const Size.square(1));
@@ -69,62 +49,7 @@ class _LiveRoomPageState extends State<LiveRoomPage> {
       body: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              DragTarget<int>(
-                builder: (context, candidateData, rejectedData) {
-                  return Container(
-                    width: 200,
-                    height: 200,
-                    color: candidateData.isNotEmpty ? Colors.green : Colors.red,
-                    child: Center(
-                      child: Text(
-                        candidateData.isNotEmpty
-                            ? 'Hovering: ${candidateData.first}' // Access the first item in the list
-                            : 'Drag an item here!',
-                        style: TextStyle(color: Colors.white),
-                      ),
-                    ),
-                  );
-                },
-                onWillAcceptWithDetails: (data) {
-                  print('Will accept: $data');
-                  return true; // Indicate whether to accept the draggable item
-                },
-                onAcceptWithDetails: (data) {
-                  voteYes();
-                  print('Accepted: $data');
-                },
-              ),
-              SongCardList(songCards: _songCards, onAdd: _addSong),
-              DragTarget<int>(
-                builder: (context, candidateData, rejectedData) {
-                  return Container(
-                    width: 200,
-                    height: 200,
-                    color: candidateData.isNotEmpty ? Colors.green : Colors.red,
-                    child: Center(
-                      child: Text(
-                        candidateData.isNotEmpty
-                            ? 'Hovering: ${candidateData.first}' // Access the first item in the list
-                            : 'Drag an item here!',
-                        style: TextStyle(color: Colors.white),
-                      ),
-                    ),
-                  );
-                },
-                onWillAcceptWithDetails: (data) {
-                  print('Will accept: $data');
-                  return true; // Indicate whether to accept the draggable item
-                },
-                onAcceptWithDetails: (data) {
-                  voteNo();
-                  print('Accepted: $data');
-                },
-              ),
-            ],
-          ),
+          Voting(initSongCards: []),
           Row(
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: [
